fix: 카테고리필터에 따라 필터된 음식점 목록 보여주도록 구현
