+++ src/components/MainContent/CategoryFilter.jsx
@@ -1,6 +1,6 @@
 import styles from "./CategoryFilter.module.css";
 
-function CategoryFilter() {
+function CategoryFilter({ category, onChangeCategory }) {
   return (
     <section className={styles["restaurant-filter-container"]}>
       <select
@@ -8,6 +8,8 @@ function CategoryFilter() {
         id="category-filter"
         className={styles["restaurant-filter"]}
         aria-label="음식점 카테고리 필터"
+        value={category}
+        onChange={(event) => onChangeCategory(event.target.value)}
       >
         <option value="전체">전체</option>
         <option value="한식">한식</option>
